[BUGFIX] Always encode full maximum speed with edges
GraphHopper by default uses a factor of 0.9 for the maximum speed of a
road, presumably to make the number more realistic given the traffic
conditions, slowdowns for turns etc.

We try to simulate these effects in RoadHopper, so we must use the full
speed instead. As it is not guaranteed that the factor will always be
applied and we cannot tell in hindsight if it was, we remove it
altogether. After a reimport of the OSM dataset, the full speed will now
be used instead.